ugc_service/requirements.txt
fastapi==0.85.1
fastapi-cache2==0.1.9
python-dotenv==0.21.0
orjson==3.8.0
backoff==2.2.1
